managed to get totally sidetracked with lua

+++ scripts/copy_shaders.py
@@ -3,5 +3,12 @@ import os
 
 in_dir = os.path.join(os.path.dirname(__file__), '..', 'shaders')
 out_dir = os.path.join(os.path.dirname(__file__), '..', 'bin', 'shaders')
-shutil.rmtree(out_dir)
+if(os.path.exists(out_dir)):
+    shutil.rmtree(out_dir)
+shutil.copytree(in_dir, out_dir)
+
+in_dir = os.path.join(os.path.dirname(__file__), '..', 'sandbox')
+out_dir = os.path.join(os.path.dirname(__file__), '..', 'bin', 'sandbox')
+if(os.path.exists(out_dir)):
+    shutil.rmtree(out_dir)
 shutil.copytree(in_dir, out_dir)
